+++ test/vmtester/tests.cpp
@@ -5,8 +5,7 @@
 #include "../../examples/example_host.h"
 #include "vmtester.hpp"
 
-#include <evmc/helpers.h>
-#include <evmc/helpers.hpp>
+#include <evmc/evmc.hpp>
 
 #include <array>
 #include <cstring>
@@ -79,7 +78,7 @@ TEST_F(evmc_vm_test, execute_call)
         read_buffer(result.output_data, result.output_size);
     }
 
-    EXPECT_TRUE(is_zero(result.create_address));
+    EXPECT_TRUE(evmc::is_zero(result.create_address));
 
     if (result.release)
         result.release(&result);
@@ -115,7 +114,7 @@ TEST_F(evmc_vm_test, execute_create)
     }
 
     // The VM will never provide the create address.
-    EXPECT_TRUE(is_zero(result.create_address));
+    EXPECT_TRUE(evmc::is_zero(result.create_address));
 
     if (result.release)
         result.release(&result);