+++ CHANGELOG.md
@@ -14,6 +14,10 @@
   Full support for 32-bit architectures has been added.
 - Added: [[#341](https://github.com/ethereum/evmc/pull/341)]
   Support for moving `evmc::vm` objects in C++ API.
+- Added: [[#357](https://github.com/ethereum/evmc/pull/357)]
+  The basic types `address` and `bytes32` have received their C++ wrappers 
+  to assure they are always initialized. They also have convenient operator
+  overloadings for comparison and usage as keys in standard containers.
 - Changed: [[#293](https://github.com/ethereum/evmc/pull/293)]
   In C++ API `evmc::result::raw()` renamed to `evmc::result::release_raw()`.
 - Changed: [[#311](https://github.com/ethereum/evmc/pull/311)]